Define `Consts` and `Values`
