+++ library/src/scala/quoted/Const.scala
@@ -26,24 +26,4 @@ object Const {
     rec(expr.unseal)
   }
 
-  /** Matches literal sequence of literal constant value expressions and return a sequence of values.
-   *
-   *  Usage:
-   *  ```scala
-   *  inline def sum(args: Int*): Int = ${ sumExpr('args) }
-   *  def sumExpr(argsExpr: Expr[Seq[Int]])(usingusing QuoteContext): Expr[Int] = argsExpr match
-   *    case Varargs(Const(args)) =>
-   *      // args: Seq[Int]
-   *      ...
-   *  }
-   *  ```
-   */
-  def unapply[T](exprs: Seq[Expr[T]])(using qctx: QuoteContext): Option[Seq[T]] =
-    exprs.foldRight(Option(List.empty[T])) { (elem, acc) =>
-      (elem, acc) match {
-        case (Const(value), Some(lst)) => Some(value :: lst)
-        case (_, _) => None
-      }
-    }
-
 }
